Prevent possible UUID-only recipient creations.

+++ src/org/thoughtcrime/securesms/crypto/storage/TextSecureSessionStore.java
@@ -15,6 +15,8 @@ import org.whispersystems.libsignal.state.SessionRecord;
 import org.whispersystems.libsignal.state.SessionStore;
 import org.whispersystems.signalservice.api.util.UuidUtil;
 
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.logging.Logger;
 
@@ -72,37 +74,52 @@ public class TextSecureSessionStore implements SessionStore {
   @Override
   public void deleteSession(SignalProtocolAddress address) {
     synchronized (FILE_LOCK) {
-      RecipientId recipientId = Recipient.external(context, address.getName()).getId();
-      DatabaseFactory.getSessionDatabase(context).delete(recipientId, address.getDeviceId());
+      if (DatabaseFactory.getRecipientDatabase(context).containsPhoneOrUuid(address.getName())) {
+        RecipientId recipientId = Recipient.external(context, address.getName()).getId();
+        DatabaseFactory.getSessionDatabase(context).delete(recipientId, address.getDeviceId());
+      } else {
+        Log.w(TAG, "Tried to delete session for " + address.toString() + ", but none existed!");
+      }
     }
   }
 
   @Override
   public void deleteAllSessions(String name) {
     synchronized (FILE_LOCK) {
-      RecipientId recipientId = Recipient.external(context, name).getId();
-      DatabaseFactory.getSessionDatabase(context).deleteAllFor(recipientId);
+      if (DatabaseFactory.getRecipientDatabase(context).containsPhoneOrUuid(name)) {
+        RecipientId recipientId = Recipient.external(context, name).getId();
+        DatabaseFactory.getSessionDatabase(context).deleteAllFor(recipientId);
+      }
     }
   }
 
   @Override
   public List<Integer> getSubDeviceSessions(String name) {
     synchronized (FILE_LOCK) {
-      RecipientId recipientId = Recipient.external(context, name).getId();
-      return DatabaseFactory.getSessionDatabase(context).getSubDevices(recipientId);
+      if (DatabaseFactory.getRecipientDatabase(context).containsPhoneOrUuid(name)) {
+        RecipientId recipientId = Recipient.external(context, name).getId();
+        return DatabaseFactory.getSessionDatabase(context).getSubDevices(recipientId);
+      } else {
+        Log.w(TAG, "Tried to get sub device sessions for " + name + ", but none existed!");
+        return Collections.emptyList();
+      }
     }
   }
 
   public void archiveSiblingSessions(@NonNull SignalProtocolAddress address) {
     synchronized (FILE_LOCK) {
-      RecipientId                      recipientId = Recipient.external(context, address.getName()).getId();
-      List<SessionDatabase.SessionRow> sessions    = DatabaseFactory.getSessionDatabase(context).getAllFor(recipientId);
+      if (DatabaseFactory.getRecipientDatabase(context).containsPhoneOrUuid(address.getName())) {
+        RecipientId                      recipientId = Recipient.external(context, address.getName()).getId();
+        List<SessionDatabase.SessionRow> sessions    = DatabaseFactory.getSessionDatabase(context).getAllFor(recipientId);
 
-      for (SessionDatabase.SessionRow row : sessions) {
-        if (row.getDeviceId() != address.getDeviceId()) {
-          row.getRecord().archiveCurrentState();
-          storeSession(new SignalProtocolAddress(Recipient.resolved(row.getRecipientId()).requireServiceId(), row.getDeviceId()), row.getRecord());
+        for (SessionDatabase.SessionRow row : sessions) {
+          if (row.getDeviceId() != address.getDeviceId()) {
+            row.getRecord().archiveCurrentState();
+            storeSession(new SignalProtocolAddress(Recipient.resolved(row.getRecipientId()).requireServiceId(), row.getDeviceId()), row.getRecord());
+          }
         }
+      } else {
+        Log.w(TAG, "Tried to archive sibling sessions for " + address.toString() + ", but none existed!");
       }
     }
   }
